refactor(f1-types): extract TeamID enums into team_id submodule with factory functions
Move all TeamID* enum classes from common.py into a new lib/f1_types/team_id.py
module. Add get_team_id(value, packet_format) factory and get_team_id_class(packet_format)
helper to centralise the if/elif dispatch that was duplicated across every parser.

- common.py re-exports the classes for backwards compatibility
- packet_4, packet_9, packet_14 _cast_enums each reduced to a single call
- test files use get_team_id_class() instead of per-format branching
- TeamID26 entries added to GENERIC_TEAMS in packet_4 (previously missing)
- TeamID26 enum class now populated from the official F1 25: 2026 Season Pack spec
  (replaces the temporary TeamID25 alias)
